move result into ts/result
