+++ ts/BUILD.bazel
@@ -5,6 +5,9 @@ package(default_visibility = ["//:__subpackages__"])
 
 ts_project(
     name = "ts",
+    deps = [
+        "//ts/result",
+    ],
 )
 
 bazel_lint(
