+++ ts/result/BUILD.bazel
@@ -1,0 +1,15 @@
+load("//bzl:rules.bzl", "bazel_lint")
+load("//ts:rules.bzl", "ts_project")
+
+ts_project(
+    name = "result",
+    srcs = ["result.ts"],
+    visibility = [
+        "//:__subpackages__",
+    ],
+)
+
+bazel_lint(
+    name = "bazel_lint",
+    srcs = ["BUILD.bazel"],
+)
